renaming the file of reporting framework
